+++ README.md
@@ -116,7 +116,7 @@ After you're finished please remove all the comments and instructions!
 
 > Add a link to your deployed project.
 
-- [Live Demo Link](https://donmark2k.github.io/todo_list/)
+- [Live Demo Link](https://donmark2k.github.io/todo_list/dist/)
 
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
