+++ src/modules/manageBooks.js
@@ -1,0 +1,8 @@
+const array = [];
+export default class Tasks {
+  constructor(description) {
+    this.description = description;
+    this.index = array.length + 1;
+    this.completed = false;
+  }
+}
